[core, ios] Introduce revalidation state API on default file source
This adds an API to default file source so that it can be put in a
state where requests are not checked for invalidation in sqlite
(or sent over the network). The iOS map view has also been updated
to use this new API when it knows that the host application is
going into the background or will otherwise be in a state where
it is not visible by the user.

This optimization means that the SDK will not make requests for data
that the user cannot see or hit edge cases on some OSs where resources
(i.e. sqlite) are not available when the host app is not visible to the
user.